+++ packages/hooks/src/useAssert/index.ts
@@ -1,1 +1,0 @@
-export { useAssert } from "./useAssert";